% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_operations.R
\name{delete_alarms}
\alias{delete_alarms}
\title{Deletes the specified alarms}
\usage{
delete_alarms(AlarmNames)
}
\arguments{
\item{AlarmNames}{The alarms to be deleted.}
}
\description{
Deletes the specified alarms. In the event of an error, no alarms are deleted.
}
\examples{

}
