% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_operations.R
\name{delete_dashboards}
\alias{delete_dashboards}
\title{Deletes all dashboards that you specify}
\usage{
delete_dashboards(DashboardNames)
}
\arguments{
\item{DashboardNames}{[required] The dashboards to be deleted. This parameter is required.}
}
\description{
Deletes all dashboards that you specify. You may specify up to 100 dashboards to delete. If there is an error during this call, no dashboards are deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_dashboards(
  DashboardNames = list(
    "string"
  )
)
}
}

