% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_operations.R
\name{describe_alarm_history}
\alias{describe_alarm_history}
\title{Retrieves the history for the specified alarm}
\usage{
describe_alarm_history(
  AlarmName = "string",
  HistoryItemType = "ConfigurationUpdate"|"StateUpdate"|"Action",
  StartDate = as.POSIXct("2015-01-01"),
  EndDate = as.POSIXct("2015-01-01"),
  MaxRecords = 123,
  NextToken = "string"
)
}
\arguments{
\item{AlarmName}{The name of the alarm.}

\item{HistoryItemType}{The type of alarm histories to retrieve.}

\item{StartDate}{The starting date to retrieve alarm history.}

\item{EndDate}{The ending date to retrieve alarm history.}

\item{MaxRecords}{The maximum number of alarm history records to retrieve.}

\item{NextToken}{The token returned by a previous call to indicate that there is more data available.}
}
\description{
Retrieves the history for the specified alarm. You can filter the results by date range or item type. If an alarm name is not specified, the histories for all alarms are returned.
}
\details{
CloudWatch retains the history of an alarm even if you delete the alarm.
}
\examples{

}
