% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_operations.R
\name{disable_alarm_actions}
\alias{disable_alarm_actions}
\title{Disables the actions for the specified alarms}
\usage{
disable_alarm_actions(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The names of the alarms.}
}
\description{
Disables the actions for the specified alarms. When an alarm's actions are disabled, the alarm actions do not execute when the alarm state changes.
}
\section{Accepted Parameters}{
\preformatted{disable_alarm_actions(
  AlarmNames = list(
    "string"
  )
)
}
}

