% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_operations.R
\name{enable_alarm_actions}
\alias{enable_alarm_actions}
\title{Enables the actions for the specified alarms}
\usage{
enable_alarm_actions(AlarmNames)
}
\arguments{
\item{AlarmNames}{[required] The names of the alarms.}
}
\description{
Enables the actions for the specified alarms.
}
\section{Accepted Parameters}{
\preformatted{enable_alarm_actions(
  AlarmNames = list(
    "string"
  )
)
}
}

