% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_package.R
\docType{package}
\name{paws.cloudwatch-package}
\alias{paws.cloudwatch}
\alias{paws.cloudwatch-package}
\title{paws.cloudwatch: Amazon CloudWatch}
\description{
Amazon CloudWatch monitors your Amazon Web Services
(AWS) resources and the applications you run on AWS in real time. You
can use CloudWatch to collect and track metrics, which are the
variables you want to measure for your resources and applications.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
