% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatch_operations.R
\name{put_metric_alarm}
\alias{put_metric_alarm}
\title{Creates or updates an alarm and associates it with the specified metric}
\usage{
put_metric_alarm(AlarmName, MetricName, Namespace, Period, EvaluationPeriods,
  Threshold, ComparisonOperator, AlarmDescription = NULL,
  ActionsEnabled = NULL, OKActions = NULL, AlarmActions = NULL,
  InsufficientDataActions = NULL, Statistic = NULL,
  ExtendedStatistic = NULL, Dimensions = NULL, Unit = NULL,
  DatapointsToAlarm = NULL, TreatMissingData = NULL,
  EvaluateLowSampleCountPercentile = NULL)
}
\arguments{
\item{AlarmName}{The name for the alarm. This name must be unique within the AWS account.}

\item{MetricName}{The name for the metric associated with the alarm.}

\item{Namespace}{The namespace for the metric associated with the alarm.}

\item{Period}{The period, in seconds, over which the specified statistic is applied. Valid values are 10, 30, and any multiple of 60.

Be sure to specify 10 or 30 only for metrics that are stored by a \code{PutMetricData} call with a \code{StorageResolution} of 1. If you specify a period of 10 or 30 for a metric that does not have sub-minute resolution, the alarm still attempts to gather data at the period rate that you specify. In this case, it does not receive data for the attempts that do not correspond to a one-minute data resolution, and the alarm may often lapse into INSUFFICENT\_DATA status. Specifying 10 or 30 also sets this alarm as a high-resolution alarm, which has a higher charge than other alarms. For more information about pricing, see \href{https://aws.amazon.com/cloudwatch/pricing/}{Amazon CloudWatch Pricing}.

An alarm's total current evaluation period can be no longer than one day, so \code{Period} multiplied by \code{EvaluationPeriods} cannot be more than 86,400 seconds.}

\item{EvaluationPeriods}{The number of periods over which data is compared to the specified threshold. If you are setting an alarm which requires that a number of consecutive data points be breaching to trigger the alarm, this value specifies that number. If you are setting an "M out of N" alarm, this value is the N.

An alarm's total current evaluation period can be no longer than one day, so this number multiplied by \code{Period} cannot be more than 86,400 seconds.}

\item{Threshold}{The value against which the specified statistic is compared.}

\item{ComparisonOperator}{The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.}

\item{AlarmDescription}{The description for the alarm.}

\item{ActionsEnabled}{Indicates whether actions should be executed during any changes to the alarm state.}

\item{OKActions}{The actions to execute when this alarm transitions to an \code{OK} state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: \code{arn:aws:automate:region:ec2:stopregion:ec2:stop} | \code{arn:aws:automate:region:ec2:terminateregion:ec2:terminate} | \code{arn:aws:automate:region:ec2:recoverregion:ec2:recover} | \code{arn:aws:sns:region:account-id:sns-topic-name region:account-id:sns-topic-name } | \code{arn:aws:autoscaling:region:account-id:scalingPolicy:policy-idautoScalingGroupName/group-friendly-name:policyName/policy-friendly-name region:account-id:scalingPolicy:policy-idautoScalingGroupName/group-friendly-name:policyName/policy-friendly-name }

Valid Values (for use with IAM roles): \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Stop/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Stop/1.0} | \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Terminate/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Terminate/1.0} | \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Reboot/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Reboot/1.0}}

\item{AlarmActions}{The actions to execute when this alarm transitions to the \code{ALARM} state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: \code{arn:aws:automate:region:ec2:stopregion:ec2:stop} | \code{arn:aws:automate:region:ec2:terminateregion:ec2:terminate} | \code{arn:aws:automate:region:ec2:recoverregion:ec2:recover} | \code{arn:aws:sns:region:account-id:sns-topic-name region:account-id:sns-topic-name } | \code{arn:aws:autoscaling:region:account-id:scalingPolicy:policy-idautoScalingGroupName/group-friendly-name:policyName/policy-friendly-name region:account-id:scalingPolicy:policy-idautoScalingGroupName/group-friendly-name:policyName/policy-friendly-name }

Valid Values (for use with IAM roles): \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Stop/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Stop/1.0} | \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Terminate/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Terminate/1.0} | \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Reboot/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Reboot/1.0}}

\item{InsufficientDataActions}{The actions to execute when this alarm transitions to the \code{INSUFFICIENT_DATA} state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: \code{arn:aws:automate:region:ec2:stopregion:ec2:stop} | \code{arn:aws:automate:region:ec2:terminateregion:ec2:terminate} | \code{arn:aws:automate:region:ec2:recoverregion:ec2:recover} | \code{arn:aws:sns:region:account-id:sns-topic-name region:account-id:sns-topic-name } | \code{arn:aws:autoscaling:region:account-id:scalingPolicy:policy-idautoScalingGroupName/group-friendly-name:policyName/policy-friendly-name region:account-id:scalingPolicy:policy-idautoScalingGroupName/group-friendly-name:policyName/policy-friendly-name }

Valid Values (for use with IAM roles): \code{>arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Stop/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Stop/1.0} | \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Terminate/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Terminate/1.0} | \code{arn:aws:swf:region:account-id:action/actions/AWS_EC2.InstanceId.Reboot/1.0region:account-id:action/actions/AWS_EC2.InstanceId.Reboot/1.0}}

\item{Statistic}{The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use \code{ExtendedStatistic}. When you call \code{PutMetricAlarm}, you must specify either \code{Statistic} or \code{ExtendedStatistic,} but not both.}

\item{ExtendedStatistic}{The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100. When you call \code{PutMetricAlarm}, you must specify either \code{Statistic} or \code{ExtendedStatistic,} but not both.}

\item{Dimensions}{The dimensions for the metric associated with the alarm.}

\item{Unit}{The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately.

If you specify a unit, you must use a unit that is appropriate for the metric. Otherwise, the CloudWatch alarm can get stuck in the \code{INSUFFICIENT DATA} state.}

\item{DatapointsToAlarm}{The number of datapoints that must be breaching to trigger the alarm. This is used only if you are setting an "M out of N" alarm. In that case, this value is the M. For more information, see \href{http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/AlarmThatSendsEmail.html#alarm-evaluation}{Evaluating an Alarm} in the \emph{Amazon CloudWatch User Guide}.}

\item{TreatMissingData}{Sets how this alarm is to handle missing data points. If \code{TreatMissingData} is omitted, the default behavior of \code{missing} is used. For more information, see \href{http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/AlarmThatSendsEmail.html#alarms-and-missing-data}{Configuring How CloudWatch Alarms Treats Missing Data}.

Valid Values: \code{breaching | notBreaching | ignore | missing}}

\item{EvaluateLowSampleCountPercentile}{Used only for alarms based on percentiles. If you specify \code{ignore}, the alarm state does not change during periods with too few data points to be statistically significant. If you specify \code{evaluate} or omit this parameter, the alarm is always evaluated and possibly changes state no matter how many data points are available. For more information, see \href{http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/AlarmThatSendsEmail.html#percentiles-with-low-samples}{Percentile-Based CloudWatch Alarms and Low Data Samples}.

Valid Values: \code{evaluate | ignore}}
}
\description{
Creates or updates an alarm and associates it with the specified metric. Optionally, this operation can associate one or more Amazon SNS resources with the alarm.
}
\details{
When this operation creates an alarm, the alarm state is immediately set to \code{INSUFFICIENT_DATA}. The alarm is evaluated and its state is set appropriately. Any actions associated with the state are then executed.

When you update an existing alarm, its state is left unchanged, but the update completely overwrites the previous configuration of the alarm.

If you are an IAM user, you must have Amazon EC2 permissions for some operations:
\itemize{
\item \code{iam:CreateServiceLinkedRole} for all alarms with EC2 actions
\item \code{ec2:DescribeInstanceStatus} and \code{ec2:DescribeInstances} for all alarms on EC2 instance status metrics
\item \code{ec2:StopInstances} for alarms with stop actions
\item \code{ec2:TerminateInstances} for alarms with terminate actions
\item \code{ec2:DescribeInstanceRecoveryAttribute} and \code{ec2:RecoverInstances} for alarms with recover actions
}

If you have read/write permissions for Amazon CloudWatch but not for Amazon EC2, you can still create an alarm, but the stop or terminate actions are not performed. However, if you are later granted the required permissions, the alarm actions that you created earlier are performed.

If you are using an IAM role (for example, an EC2 instance profile), you cannot stop or terminate the instance using alarm actions. However, you can still see the alarm state and perform any other actions such as Amazon SNS notifications or Auto Scaling policies.

If you are using temporary security credentials granted using AWS STS, you cannot stop or terminate an EC2 instance using alarm actions.

The first time you create an alarm in the AWS Management Console, the CLI, or by using the PutMetricAlarm API, CloudWatch creates the necessary service-linked role for you. The service-linked role is called \code{AWSServiceRoleForCloudWatchEvents}. For more information about service-linked roles, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_terms-and-concepts.html#iam-term-service-linked-role}{AWS service-linked role}.
}
\examples{

}
